+++ pages/rank.tsx
@@ -1,4 +1,4 @@
-import ModeSelect from 'components/mode/ModeSelect';
+import ModeSeasonProvider from 'components/mode/ModeSeasonProvider';
 import MyRank from 'components/rank/MyRank';
 import RankList from 'components/rank/RankList';
 import styles from 'styles/RankList.module.scss';
@@ -8,9 +8,9 @@ export default function Rank() {
     <div className={styles.pageWrap}>
       <h1 className={styles.title}>Ranking</h1>
       <MyRank />
-      <ModeSelect>
+      <ModeSeasonProvider>
         <RankList />
-      </ModeSelect>
+      </ModeSeasonProvider>
     </div>
   );
 }
