--- pages/test.tsx
@@ -1,11 +1,0 @@
-import ModeSelect from 'components/mode/ModeSelect';
-
-export default function test() {
-  return (
-    <div>
-      <ModeSelect>
-        <div>test</div>
-      </ModeSelect>
-    </div>
-  );
-}
