+++ pages/users/detail.tsx
@@ -5,9 +5,8 @@ import Chart from 'components/user/Chart';
 import Profile from 'components/user/Profile';
 import GameResult from 'components/game/GameResult';
 import styles from 'styles/user/user.module.scss';
-
 import { useEffect } from 'react';
-import ModeSelect from 'components/mode/ModeSelect';
+import ModeSeasonProvider from 'components/mode/ModeSeasonProvider';
 
 export default function user() {
   const setModalInfo = useSetRecoilState(modalState);
@@ -23,9 +22,9 @@ export default function user() {
       {typeof intraId === 'string' && (
         <>
           <h1 className={styles.title}>{intraId}</h1>
-          <ModeSelect>
+          <ModeSeasonProvider>
             <UserInfo intraId={intraId} />
-          </ModeSelect>
+          </ModeSeasonProvider>
         </>
       )}
     </div>
